requirements.txt
Click==7.0
boto3==1.9.215
SQLAlchemy==1.3.8
mysqlclient==1.4.5
python-dotenv==0.10.3
alembic==1.0.10
shortuuid==0.5.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Bcrypt==0.7.1
Flask-Script==2.0.6
Flask-Injector==0.12.0
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
PyJWT==1.7.1
Werkzeug==0.15.6